kasa_electricity_recorder_function/requirements.txt
# For Tp-link Kasa
tplink-cloud-api==1.0.2
myb-google-sheets-api>=1.0.0
